cloud_storage: move tests for partition_manifest to a separate file
