+++ src/v/cloud_storage/tests/CMakeLists.txt
@@ -1,21 +1,18 @@
-
 rp_test(
   UNIT_TEST
   BINARY_NAME test_cloud_storage
-  SOURCES 
+  SOURCES
     directory_walker_test.cc
-    manifest_test.cc 
-    s3_imposter.cc 
-    remote_test.cc 
-    cache_test.cc  
-    offset_translation_layer_test.cc 
-    remote_segment_test.cc 
+    manifest_test.cc
+    partition_manifest_test.cc
+    s3_imposter.cc
+    remote_test.cc
+    cache_test.cc
+    offset_translation_layer_test.cc
+    remote_segment_test.cc
     remote_partition_test.cc
   DEFINITIONS BOOST_TEST_DYN_LINK
   LIBRARIES v::seastar_testing_main Boost::unit_test_framework v::cloud_storage v::storage_test_utils
   ARGS "-- -c 1"
   LABELS cloud_storage
 )
-
-
-
